Feature/add validations form

+++ src/Presentation/components/tables/TableInformation.tsx
@@ -3,6 +3,7 @@ import { IColumnsDataTable } from "../../interfaces/interfaces";
 import { DataGrid } from "@mui/x-data-grid";
 import { useDispatch } from "react-redux";
 import { Dispatch } from "redux";
+import Swal from "sweetalert2";
 
 interface ITableInformationProps {
   categories?: any[] | [];
@@ -41,7 +42,24 @@ export const TableInformation = ({
 
   const handleDelete = (params: any) => {
     if (deleteCategory) {
-      dispatch(deleteCategory(params.row.id));
+      Swal.fire({
+        title: "¿Estás seguro?",
+        text: "¡No podrás deshacer esta acción!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "!Sí, eliminar!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          dispatch(deleteCategory(params.row.id));
+          Swal.fire(
+            "¡Eliminado!",
+            "Ha sido eliminado correctamente.",
+            "success"
+          );
+        }
+      });
     }
   };
